存放区域、登录、操作日志、管理员复核等

+++ pawn/src/main/java/cn/cnplay/demo/service/HomeService.java
@@ -36,7 +36,7 @@ public class HomeService {
 		if(StringUtils.isBlank(user.getPassword())){
 			return ResultVOUtil.error("密码不能为空");
 		}
-		List<User> list = userRepository.findByUsername(user.getUsername());
+		List<User> list = userRepository.findByUsernameAndDeleted(user.getUsername(),false);
 		if(list.size()==0){
 			return ResultVOUtil.error("无此用户");
 		}
@@ -49,6 +49,7 @@ public class HomeService {
 		if(StringUtils.equals(checksum, poPswd)){
 			user = userPO;
 			String token = jwtHelper.createToken(user);
+			tokenLogRepository.deleteUserToken(user.getId());
 			UserTokenLog tokenLog = new UserTokenLog();
 			tokenLog.setJwtToken(token);
 			tokenLog.setUser(user);
@@ -59,6 +60,8 @@ public class HomeService {
 		}
 	}
 	
+	
+	
 	@Transactional
 	public void logout(String token){
 		if(StringUtils.isBlank(token)){
@@ -67,7 +70,36 @@ public class HomeService {
 		User user = jwtHelper.verifyToken(token);
 		if(user!=null){
 			UserTokenLog tokenLog = tokenLogRepository.getUserToken(user.getId());
-			tokenLogRepository.delete(tokenLog);
+			if(StringUtils.equals(token, tokenLog.getJwtToken())){
+				tokenLogRepository.delete(tokenLog);
+			}
+		}
+	}
+	
+	public ResultVo<?> audit(User loginUser,User auditUser){
+		List<User> list = userRepository.findByUsernameAndDeleted(auditUser.getUsername(),false);
+		if(list.size()==0){
+			return ResultVOUtil.error("用户不存在");
+		}
+		User userPO = list.get(0);
+		if(!StringUtils.equals("admin", userPO.getUserType())){
+			return ResultVOUtil.error("管理员权限才能进行复核");
+		}
+		if(loginUser.getDepartment()==null||auditUser.getDepartment()==null){
+			return ResultVOUtil.error("登录账号与复核账号存在机构信息为空");
+		}
+		if(!StringUtils.equals(loginUser.getDepartment().getId(), auditUser.getDepartment().getId())){
+			return ResultVOUtil.error("只有同一机构才能进行复核");
+		}
+		String voPswd = auditUser.getPassword();
+		String poPswd = userPO.getPassword();
+		String saltStr = StringUtils.substring(poPswd, 0, 16);
+		poPswd = StringUtils.substring(poPswd, 16);
+		String checksum = Sha1Util.encrypt(voPswd, saltStr);
+		if(StringUtils.equals(checksum, poPswd)){
+			return ResultVOUtil.success(userPO.getId());
+		}else{
+			return ResultVOUtil.error("密码错误");
 		}
 	}
 }
